Make compatible with Marshmallow >= 3.0.0b7
Marshmallow has changed the return value of schema.dump(), as per the note in
the API docs:

    http://marshmallow.readthedocs.io/en/latest/api_reference.html#marshmallow.Schema.dump

"Changed in version 3.0.0b7: This method returns the serialized data rather than
a (data, errors) duple. A ValidationError is raised if obj is invalid."

This change just checks the return value of dump().  If it's a dict, it returns
that as output, otherwise, it returns the .data attribute as before.