Skip static rules
When using Flask 2x the send_static_file function is a lambda and
crashes the extraction of the rules_by_endpoint function. There is
currently a PR that has a fix but it's been stale so I'm adding this as
a possible alternate fix.

The other fix is: https://github.com/jmcarp/flask-apispec/pull/230